[NEWS] Add new AQ annotation and improve trio-dnm calling
The original trio-dnm calling model used genotype likelihoods (PLs) as the
input for calling. However, that is flawed because PLs make assumptions which
are unsuitable for de novo calling: PL(RR) can become bigger than PL(RA) even
when the ALT allele is present in the parents. Note that this is true also
for other programs such as DeNovoGear which rely on the same samtools calculation.

The use of AQ is the new default and the recommended workflow is

    bcftools mpileup -a AD,AQ -f ref.fa -Ou proband.bam father.bam mother.bam | \
    bcftools call -mv -Ou | \
    bcftools +trio-dnm -p proband,father,mother -Oz -o output.vcf.gz

The old behavior can be evoked by running +trio-dnm with the --no-AQ option.